Move navigation and run code keybinds logic out out of Question components and into ProblemView
